backend/requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
python-dotenv
google-generativeai
pandas
scikit-learn
joblib
xgboost
lightgbm
catboost
python-multipart
